Move CMOS and RTC to clk module
