+++ src/sys/clk/mod.rs
@@ -1,0 +1,7 @@
+mod cmos;
+mod rtc;
+
+pub use cmos::CMOS;
+pub use rtc::RTC;
+
+use rtc::{Interrupt, Register, RTC_CENTURY};
