+++ src/sys/mod.rs
@@ -40,7 +40,7 @@ macro_rules! log {
 pub mod acpi;
 pub mod ata;
 pub mod clock;
-pub mod cmos;
+pub mod clk;
 pub mod console;
 pub mod cpu;
 pub mod fs;
